% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_stash.s3_stash.R
\name{delete_stash_.s3_stash}
\alias{delete_stash_.s3_stash}
\title{Delete a file/object from an AWS S3 stash}
\usage{
\method{delete_stash_}{s3_stash}(file.name, from, time.stamp, uuid, extension,
  compression, checksum, no.prompt, clean.up)
}
\arguments{
\item{file.name}{Base file name for the object}

\item{from}{Path for where the object resides. Accepts a character
string or *_stash object.}

\item{time.stamp}{TRUE or FALSE. When TRUE, files containing a time stamp
generated by save_stash, or matching the save_stash pattern, will be
deleted.}

\item{uuid}{TRUE or FALSE. When TRUE, files containing a uuid
generated by save_stash, or matching the save_stash pattern, will be
deleted.}

\item{extension}{Extension for the file to delete.}

\item{compression}{Accepts NULL for no compression, or 'gz' for gzip.}

\item{checksum}{Not being used right now.}

\item{no.prompt}{TRUE or FALSE. When FALSE, you will be promted to confirm
deletion of matching files. When TRUE, you will not be prompted. Defaults
to FALSE.}

\item{clean.up}{TRUE or FALSE. When TRUE, the directory that the file resided
in will be deleted if it is empty.}
}
\value{
List of stash objects that were successfully deleted.
}
\description{
Delete a file/object from an AWS S3 stash
}

