% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftp_stash.R
\name{ftp_stash}
\alias{ftp_stash}
\title{Object for representing a ftp file or directory}
\usage{
ftp_stash(server, path = "", time.stamp = "auto", uuid = "auto",
  extension = "auto", compression = "auto", base.file = NULL,
  is.file = FALSE, user = NULL, password = NULL, ..., simplify = TRUE)
}
\arguments{
\item{path}{A file path to a local directory or file.}

\item{compression}{Default to auto, if the file ends it '.gz' it will identify gzip.}
}
\value{
A ftp_stash object.
}
\description{
For extension and compression, default to 'auto'. Check if available compression
is applied and extract it. If you do not want any compression to be identified
then set to NULL If the compression extension is named differently then was
rstash would create, pass a named vector, the name being the rstash designation
and the value being the compression extension of the existing file. For
example compression = c(gz = 'gzip') or c(gz = NA) if the compression is gzip
but there is no extension on the file.
}
\details{
If this is not a file then nothing is processed, for 'auto', the
}

