% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stash_.s3_stash.R
\name{get_stash_.s3_stash}
\alias{get_stash_.s3_stash}
\title{Get/Load a stashed object from a AWS S3 stash}
\usage{
\method{get_stash_}{s3_stash}(file.name, from, time.stamp, uuid, extension,
  compression, checksum, read.fn, ...)
}
\arguments{
\item{file.name}{Base file name for the object}

\item{from}{Path for where the object resides. Accepts a character
string or *_stash object.}

\item{time.stamp}{TRUE or FALSE. When TRUE, files containing a time stamp
generated by save_stash, or matching the save_stash pattern, will be
loaded.}

\item{uuid}{TRUE or FALSE. When TRUE, files containing a uuid
generated by save_stash, or matching the save_stash pattern, will be
loaded.}

\item{extension}{Extension for the file to load.}

\item{compression}{Accepts NULL for no compression, or 'gz' for gzip.}

\item{checksum}{Not being used right now.}

\item{read.fn}{Any function where the first argument is the object and the
second is the file path. Will be used to read the files into R.}

\item{...}{Arguments to pass to \code{read.fn}.}
}
\value{
List of loaded stash objects.
}
\description{
Get/Load a stashed object from a AWS S3 stash
}

