% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_dirs.R
\name{list_dirs}
\alias{list_dirs}
\title{List the Directory in a Stash or Directory/Folder}
\usage{
list_dirs(x = ".", full.names = TRUE, recursive = TRUE, as.stash = TRUE,
  simplify = TRUE)
}
\description{
List the Directory in a Stash or Directory/Folder
}

