% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files.R
\name{list_files}
\alias{list_files}
\title{List the Files in a Stash or Directory/Folder}
\usage{
list_files(x = ".", pattern = NULL, all.files = FALSE,
  full.names = FALSE, recursive = FALSE, ignore.case = FALSE,
  include.dirs = !recursive, no.. = FALSE, match.file = TRUE,
  as.stash = TRUE, simplify = TRUE)
}
\arguments{
\item{x}{A \code{character} vector or a \code{\link{stash}}.}

\item{pattern}{Optional \code{\link[base]{regex}} (regular expression).
Only file names which match the regular expression will be returned.}

\item{all.files}{Logical. If \code{FALSE}, only the names of visible files
are returned. If TRUE, all file names will be returned. \strong{Not
applicable when \code{x} is a \code{\link{ftp_stash}} or
\code{\link{s3_stash}}.}}

\item{full.names}{Logical. If \code{FALSE} and \code{recursive = FALSE}, only
the file names are returned. If \code{FALSE} and \code{recursive = TRUE},
sub-directory names will be prepended to the file names. If \code{TRUE}
and \code{x} is a \code{character} vector or a
\code{\link{local_stash}}, the full directory path is prepended to the
file names to give a relative file path. If \code{TRUE} and \code{x} is a
\code{\link{ftp_stash}} or \code{\link{s3_stash}}, then a fully resolvable
URL will be returned. \strong{Not applicable when \code{as.stash = TRUE}.}}

\item{recursive}{Logical. Should the listing recurse into directories?}

\item{ignore.case}{Logical. Should pattern-matching be case-insensitive?}

\item{include.dirs}{Logical. If \code{TRUE}, then sub-directory
names be included when \code{recursive = TRUE}, no effect when
\code{recursive = FALSE} (sub-directory names will always be included).}

\item{no..}{Logical. Should both "." and ".." be excluded from non-recursive
listings? \strong{Not applicable when \code{x} is a \code{\link{ftp_stash}}
or \code{\link{s3_stash}}.}}

\item{match.file}{Logical. If \code{TRUE} and \code{x} is a file stash, then
the file name pattern will be extracted and used for \code{pattern}. Not
applicable is \code{x} is not a stash object or is a directory.}

\item{as.stash}{Logical. If \code{TRUE}, a list of stash objects will be
returned. If \code{FALSE}, a \code{character} vector is returned.}
}
\value{
If \code{as.stash = TRUE}, a list of \code{\link{stash}} objects
   will be returned, matching the stash type of \code{x}. If \code{x} is a
   \code{character} vector then a list of \code{\link{local_stash}} objects
   are returned. If \code{as.stash = FALSE}, a \code{character} vector is
   returned. If there are no matching files, the list or character vector
   will be of \code{length == 0}.
}
\description{
When \code{x} is a \code{\link[base]{character}} vector or a
\code{\link{local_stash}}, then \code{list_files} behaves as an alias for
\code{\link[base]{list.files}} where \code{x} is passed to the \code{path}
argument of \code{\link[base]{list.files}} and the remaining matching
arguments are passed as is with the default values of \code{list_files} where
applicable. (If you do not define an argument, the default value for the
argument as defined in \code{list_files} will be used and not the default as
defined in \code{\link[base]{list.files}}).
}

