% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstash.R
\name{local_stash}
\alias{local_stash}
\title{Object for representing a local file or directory}
\usage{
local_stash(path = "")
}
\arguments{
\item{path}{A file path to a local directory or file.}
}
\value{
A local_stash object.
}
\description{
Object for representing a local file or directory
}

