% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_stash.R
\name{move_stash}
\alias{move_stash}
\title{Move a file/object from one stash to another}
\usage{
move_stash(file.name, from = "", to = "", time.stamp = FALSE,
  uuid = FALSE, extension = NULL, compression = NULL, checksum = FALSE,
  clean.up = TRUE)
}
\arguments{
\item{file.name}{Base file name for the object}

\item{from}{Path for where the object currently resides. Accepts a character
string or *_stash object.}

\item{to}{Path for where the object will be moved to. Accepts a character
string or *_stash object.}

\item{time.stamp}{TRUE or FALSE. When TRUE, files containing a time stamp
generated by save_stash, or matching the save_stash pattern, will be
moved.}

\item{uuid}{TRUE or FALSE. When TRUE, files containing a uuid
generated by save_stash, or matching the save_stash pattern, will be
moved.}

\item{extension}{Extension for the file to copy.}

\item{compression}{Accepts NULL for no compression, or 'gz' for gzip.}

\item{checksum}{Not being used right now.}

\item{clean.up}{TRUE or FALSE. When TRUE, the directory that the file
was moved from will be deleted if it is empty.}
}
\value{
List of stash objects that were successfully moved to their
   destination.
}
\description{
Same functionality as \code{copy_stash} when \code{keep.stash = TRUE}.
}

