% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstash.R
\docType{package}
\name{rstash}
\alias{rstash}
\alias{rstash-package}
\title{rstash: A package for...}
\description{
The foo package provides three categories of important functions:
foo, bar and baz.
}
\section{Foo functions}{

The foo functions ...
}

