% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstash.R
\name{s3_stash}
\alias{s3_stash}
\title{Object for representing an AWS S3 file or directory}
\usage{
s3_stash(bucket, path = "")
}
\arguments{
\item{bucket}{AWS S3 bucket}

\item{path}{A "file path" to a AWS S3 directory or file. S3 doesn't
technically have folders/directories, but the object keys can be
interpreted that way.}
}
\value{
A s3_stash object.
}
\description{
Object for representing an AWS S3 file or directory
}

