% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_stash.s3_stash.R
\name{save_stash_.s3_stash}
\alias{save_stash_.s3_stash}
\title{Save an object to a AWS S3 stash}
\usage{
\method{save_stash_}{s3_stash}(x, file.name, to, time.stamp, uuid, extension,
  compression, checksum, save.fn, ...)
}
\arguments{
\item{x}{Object to save}

\item{file.name}{Base file name for the saved object}

\item{to}{Path for where the object will be saved. Accepts a character
string or *_stash object.}

\item{time.stamp}{TRUE or FALSE. When TRUE, a time stamp will be appended to
the \code{file.name}. It is recommended to use time.stamp OR uuid, not
both. Using both can cause the file name to become very long.}

\item{uuid}{TRUE or FALSE. When TRUE, a uuid will be appended to
the \code{file.name}. It is recommended to use time.stamp OR uuid, not
both. Using both can cause the file name to become very long.}

\item{extension}{Extension to append to the file.}

\item{compression}{Accepts NULL for no compression, or 'gz' for gzip.}

\item{checksum}{Not being used right now.}

\item{save.fn}{Any function where the first argument is the object and the
second is the file path.}

\item{...}{Arguments to pass to \code{save.fn}.}
}
\value{
List of stash objects that were successfully saved.
}
\description{
Save an object to a AWS S3 stash
}

