% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stash_content.R
\name{stash_content}
\alias{stash_content}
\title{List all files/objects in a stash}
\usage{
stash_content(from = "", recursive = FALSE)
}
\arguments{
\item{from}{Path for where the object resides. Accepts a character
string or *_stash object.}

\item{recursive}{Logical. Should the contents recurse into directories?}
}
\value{
List of stash objects residing in the \code{from} stash
}
\description{
List all files/objects in a stash
}

