% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stash_match.R
\name{stash_match}
\alias{stash_match}
\title{List files/objects that match parameters}
\usage{
stash_match(file.name, from = "", time.stamp = FALSE, uuid = FALSE,
  extension = NULL, compression = NULL, recursive = FALSE)
}
\arguments{
\item{file.name}{Base file name for the object}

\item{from}{Path for where the object resides. Accepts a character
string or *_stash object.}

\item{time.stamp}{TRUE or FALSE. When TRUE, files containing a time stamp
generated by save_stash, or matching the save_stash pattern, will be
matched.}

\item{uuid}{TRUE or FALSE. When TRUE, files containing a uuid
generated by save_stash, or matching the save_stash pattern, will be
matched.}

\item{extension}{Extension for the file to load.}

\item{compression}{Accepts NULL for no compression, or 'gz' for gzip.}

\item{recursive}{Logical. Should the match recurse into directories?}
}
\value{
List of stash objects that match parameters.
}
\description{
List files/objects that match parameters
}

