% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_stash.R
\name{copy_stash}
\alias{copy_stash}
\title{Copy a file/object from one stash to another}
\usage{
copy_stash(from, to, one.to = "one", keep.from = TRUE, clean.up = TRUE,
  simplify = TRUE)
}
\arguments{
\item{from}{Path for where the object currently resides. Accepts a character
string or *_stash object.}

\item{to}{Path for where the object will be copied to. Accepts a character
string or *_stash object.}

\item{keep.from}{TRUE or FALSE. When TRUE, the files/objects that were
copied from will be deleted.}

\item{clean.up}{TRUE or FALSE. Only used when \code{keep.from} equals FALSE
When TRUE, and \code{keep.from} equals FALSE, the directory that the file
was copied from will be deleted if it is empty.}

\item{file.name}{Base file name for the object}

\item{time.stamp}{TRUE or FALSE. When TRUE, files containing a time stamp
generated by save_stash, or matching the save_stash pattern, will be
copied.}

\item{uuid}{TRUE or FALSE. When TRUE, files containing a uuid
generated by save_stash, or matching the save_stash pattern, will be
copied.}

\item{extension}{Extension for the file to copy.}

\item{compression}{Accepts NULL for no compression, or 'gz' for gzip.}

\item{checksum}{Not being used right now.}
}
\value{
List of stash objects that were successfully copied to their
   destination.
}
\description{
Copy a file/object from one stash to another
}

