% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_stash.R
\name{s3_stash}
\alias{s3_stash}
\title{Object for representing an AWS S3 file or directory}
\usage{
s3_stash(bucket, path = "", time.stamp = "auto", uuid = "auto",
  extension = "auto", compression = "auto", base.file = NULL,
  is.file = FALSE, access.key.id = NULL, secret.access.key = NULL,
  region = NULL, url.style = "virtual-hosted", ..., simplify = TRUE)
}
\arguments{
\item{bucket}{AWS S3 bucket}

\item{path}{A "file path" to a AWS S3 directory or file. S3 doesn't
technically have folders/directories, but the object keys can be
interpreted that way.}

\item{url.style}{'virtual-hosted' or 'path'. see:
https://docs.aws.amazon.com/AmazonS3/latest/dev/UsingBucket.html
Not really important for rstash operations, but may be useful for other APIs}
}
\value{
A s3_stash object.
}
\description{
Object for representing an AWS S3 file or directory
}

